+++ lib/molecules/dynamic_input/inputs/zwap_check_box_picker/zwap_check_box_picker_provider.dart
@@ -7,18 +7,22 @@ class _ZwapCheckBoxPickerProvider extends ChangeNotifier {
   final Function(String, bool)? onToggleItem;
   final Function()? onClearAll;
 
+  final int minSelectedItems;
+
   _ZwapCheckBoxPickerProvider({
     required this.onToggleItem,
     required this.values,
     required this.onClearAll,
+    required this.minSelectedItems,
     List<String>? initialSelectedKeys,
   })  : this.selectedKeys = initialSelectedKeys ?? [],
         super();
 
   void toggleItem(String item) {
-    if (selectedKeys.contains(item))
+    if (selectedKeys.contains(item)) {
+      if (minSelectedItems > 0 && selectedKeys.length - 1 < minSelectedItems) return;
       selectedKeys = List.from(selectedKeys)..remove(item);
-    else
+    } else
       selectedKeys = List.from(selectedKeys)..add(item);
 
     if (onToggleItem != null) onToggleItem!(item, selectedKeys.contains(item));
@@ -28,8 +32,13 @@ class _ZwapCheckBoxPickerProvider extends ChangeNotifier {
 
   void clear() {
     if (onClearAll != null) onClearAll!();
-    
-    selectedKeys = [];
+
+    if (minSelectedItems == 0) {
+      selectedKeys = [];
+    } else {
+      while (selectedKeys.length > minSelectedItems) selectedKeys.removeLast();
+    }
+
     notifyListeners();
   }
 }